+++ code/core/unit_tests/src/BodyTest.cpp
@@ -45,17 +45,17 @@ TEST_F(BodyTest, can_build_rotation_matrix_from_states)
 {
     const StateType x = {1,2,3,4,5,6,7,8,9,10,11,12,13,1,2,3,4,5,6,7,8,9,3,5,7,13};
     const ssc::kinematics::RotationMatrix R = body->get_rot_from_ned_to(x);
-    ASSERT_DOUBLE_EQ(1-2*7*7-2*13*13, (double)R(0,0));
-    ASSERT_DOUBLE_EQ(2*5*7+2*13*3,    (double)R(1,0));
-    ASSERT_DOUBLE_EQ(2*5*13-2*7*3,    (double)R(2,0));
+    ASSERT_DOUBLE_EQ(1-2*7*7-2*13*13, R(0,0));
+    ASSERT_DOUBLE_EQ(2*5*7+2*13*3,    R(1,0));
+    ASSERT_DOUBLE_EQ(2*5*13-2*7*3,    R(2,0));
 
-    ASSERT_DOUBLE_EQ(2*5*7-2*13*3,    (double)R(0,1));
-    ASSERT_DOUBLE_EQ(1-2*5*5-2*13*13, (double)R(1,1));
-    ASSERT_DOUBLE_EQ(2*7*13+2*5*3,    (double)R(2,1));
+    ASSERT_DOUBLE_EQ(2*5*7-2*13*3,    R(0,1));
+    ASSERT_DOUBLE_EQ(1-2*5*5-2*13*13, R(1,1));
+    ASSERT_DOUBLE_EQ(2*7*13+2*5*3,    R(2,1));
 
-    ASSERT_DOUBLE_EQ(2*5*13+2*7*3,    (double)R(0,2));
-    ASSERT_DOUBLE_EQ(2*7*13-2*5*3,    (double)R(1,2));
-    ASSERT_DOUBLE_EQ(1-2*5*5-2*7*7,   (double)R(2,2));
+    ASSERT_DOUBLE_EQ(2*5*13+2*7*3,    R(0,2));
+    ASSERT_DOUBLE_EQ(2*7*13-2*5*3,    R(1,2));
+    ASSERT_DOUBLE_EQ(1-2*5*5-2*7*7,   R(2,2));
 }
 
 TEST_F(BodyTest, can_return_position_of_body_mesh_from_Body_object)
@@ -77,17 +77,17 @@ TEST_F(BodyTest, can_compute_transform_from_mesh_to_body_from_Body_object)
     ASSERT_DOUBLE_EQ(0.21, T.get_point().y());
     ASSERT_DOUBLE_EQ(33E3, T.get_point().z());
 
-    ASSERT_DOUBLE_EQ(cos(2)*cos(3),(double)T.get_rot()(0,0));
-    ASSERT_DOUBLE_EQ(sin(2)*cos(3),(double)T.get_rot()(1,0));
-    ASSERT_DOUBLE_EQ(-sin(3),      (double)T.get_rot()(2,0));
+    ASSERT_DOUBLE_EQ(cos(2)*cos(3),T.get_rot()(0,0));
+    ASSERT_DOUBLE_EQ(sin(2)*cos(3),T.get_rot()(1,0));
+    ASSERT_DOUBLE_EQ(-sin(3),      T.get_rot()(2,0));
 
-    ASSERT_DOUBLE_EQ(-sin(2)*cos(1)+cos(2)*sin(3)*sin(1),(double)T.get_rot()(0,1));
-    ASSERT_DOUBLE_EQ(cos(2)*cos(1)+sin(2)*sin(3)*sin(1), (double)T.get_rot()(1,1));
-    ASSERT_DOUBLE_EQ(cos(3)*sin(1),                      (double)T.get_rot()(2,1));
+    ASSERT_DOUBLE_EQ(-sin(2)*cos(1)+cos(2)*sin(3)*sin(1),T.get_rot()(0,1));
+    ASSERT_DOUBLE_EQ(cos(2)*cos(1)+sin(2)*sin(3)*sin(1), T.get_rot()(1,1));
+    ASSERT_DOUBLE_EQ(cos(3)*sin(1),                      T.get_rot()(2,1));
 
-    ASSERT_DOUBLE_EQ(sin(2)*sin(1)+cos(2)*cos(1)*sin(3), (double)T.get_rot()(0,2));
-    ASSERT_DOUBLE_EQ(-cos(2)*sin(1)+sin(3)*sin(2)*cos(1),(double)T.get_rot()(1,2));
-    ASSERT_DOUBLE_EQ(cos(3)*cos(1),                      (double)T.get_rot()(2,2));
+    ASSERT_DOUBLE_EQ(sin(2)*sin(1)+cos(2)*cos(1)*sin(3), T.get_rot()(0,2));
+    ASSERT_DOUBLE_EQ(-cos(2)*sin(1)+sin(3)*sin(2)*cos(1),T.get_rot()(1,2));
+    ASSERT_DOUBLE_EQ(cos(3)*cos(1),                      T.get_rot()(2,2));
 }
 
 TEST_F(BodyTest, can_get_transform_from_NED_to_body_from_states)
@@ -105,17 +105,17 @@ TEST_F(BodyTest, can_get_transform_from_NED_to_body_from_states)
     ASSERT_DOUBLE_EQ(2, T.get_point().y());
     ASSERT_DOUBLE_EQ(3, T.get_point().z());
 
-    ASSERT_DOUBLE_EQ(1-2*7*7-2*13*13, (double)T.get_rot()(0,0));
-    ASSERT_DOUBLE_EQ(2*5*7+2*13*3,    (double)T.get_rot()(1,0));
-    ASSERT_DOUBLE_EQ(2*5*13-2*7*3,    (double)T.get_rot()(2,0));
+    ASSERT_DOUBLE_EQ(1-2*7*7-2*13*13, T.get_rot()(0,0));
+    ASSERT_DOUBLE_EQ(2*5*7+2*13*3,    T.get_rot()(1,0));
+    ASSERT_DOUBLE_EQ(2*5*13-2*7*3,    T.get_rot()(2,0));
 
-    ASSERT_DOUBLE_EQ(2*5*7-2*13*3,    (double)T.get_rot()(0,1));
-    ASSERT_DOUBLE_EQ(1-2*5*5-2*13*13, (double)T.get_rot()(1,1));
-    ASSERT_DOUBLE_EQ(2*7*13+2*5*3,    (double)T.get_rot()(2,1));
+    ASSERT_DOUBLE_EQ(2*5*7-2*13*3,    T.get_rot()(0,1));
+    ASSERT_DOUBLE_EQ(1-2*5*5-2*13*13, T.get_rot()(1,1));
+    ASSERT_DOUBLE_EQ(2*7*13+2*5*3,    T.get_rot()(2,1));
 
-    ASSERT_DOUBLE_EQ(2*5*13+2*7*3,    (double)T.get_rot()(0,2));
-    ASSERT_DOUBLE_EQ(2*7*13-2*5*3,    (double)T.get_rot()(1,2));
-    ASSERT_DOUBLE_EQ(1-2*5*5-2*7*7,   (double)T.get_rot()(2,2));
+    ASSERT_DOUBLE_EQ(2*5*13+2*7*3,    T.get_rot()(0,2));
+    ASSERT_DOUBLE_EQ(2*7*13-2*5*3,    T.get_rot()(1,2));
+    ASSERT_DOUBLE_EQ(1-2*5*5-2*7*7,   T.get_rot()(2,2));
 }
 
 TEST_F(BodyTest, can_update_Kinematics_object_from_states)
@@ -132,17 +132,17 @@ TEST_F(BodyTest, can_update_Kinematics_object_from_states)
     ASSERT_DOUBLE_EQ(2, T.get_point().y());
     ASSERT_DOUBLE_EQ(3, T.get_point().z());
 
-    ASSERT_DOUBLE_EQ(1-2*7*7-2*13*13, (double)T.get_rot()(0,0));
-    ASSERT_DOUBLE_EQ(2*5*7+2*13*3,    (double)T.get_rot()(1,0));
-    ASSERT_DOUBLE_EQ(2*5*13-2*7*3,    (double)T.get_rot()(2,0));
+    ASSERT_DOUBLE_EQ(1-2*7*7-2*13*13, T.get_rot()(0,0));
+    ASSERT_DOUBLE_EQ(2*5*7+2*13*3,    T.get_rot()(1,0));
+    ASSERT_DOUBLE_EQ(2*5*13-2*7*3,    T.get_rot()(2,0));
 
-    ASSERT_DOUBLE_EQ(2*5*7-2*13*3,    (double)T.get_rot()(0,1));
-    ASSERT_DOUBLE_EQ(1-2*5*5-2*13*13, (double)T.get_rot()(1,1));
-    ASSERT_DOUBLE_EQ(2*7*13+2*5*3,    (double)T.get_rot()(2,1));
+    ASSERT_DOUBLE_EQ(2*5*7-2*13*3,    T.get_rot()(0,1));
+    ASSERT_DOUBLE_EQ(1-2*5*5-2*13*13, T.get_rot()(1,1));
+    ASSERT_DOUBLE_EQ(2*7*13+2*5*3,    T.get_rot()(2,1));
 
-    ASSERT_DOUBLE_EQ(2*5*13+2*7*3,    (double)T.get_rot()(0,2));
-    ASSERT_DOUBLE_EQ(2*7*13-2*5*3,    (double)T.get_rot()(1,2));
-    ASSERT_DOUBLE_EQ(1-2*5*5-2*7*7,   (double)T.get_rot()(2,2));
+    ASSERT_DOUBLE_EQ(2*5*13+2*7*3,    T.get_rot()(0,2));
+    ASSERT_DOUBLE_EQ(2*7*13-2*5*3,    T.get_rot()(1,2));
+    ASSERT_DOUBLE_EQ(1-2*5*5-2*7*7,   T.get_rot()(2,2));
 }
 
 TEST_F(BodyTest, can_compute_transform_from_ned_to_local_ned)
@@ -160,17 +160,17 @@ TEST_F(BodyTest, can_compute_transform_from_ned_to_local_ned)
     ASSERT_DOUBLE_EQ(2, T.get_point().y());
     ASSERT_DOUBLE_EQ(3, T.get_point().z());
 
-    ASSERT_DOUBLE_EQ(1, (double)T.get_rot()(0,0));
-    ASSERT_DOUBLE_EQ(0, (double)T.get_rot()(1,0));
-    ASSERT_DOUBLE_EQ(0, (double)T.get_rot()(2,0));
+    ASSERT_DOUBLE_EQ(1, T.get_rot()(0,0));
+    ASSERT_DOUBLE_EQ(0, T.get_rot()(1,0));
+    ASSERT_DOUBLE_EQ(0, T.get_rot()(2,0));
 
-    ASSERT_DOUBLE_EQ(0, (double)T.get_rot()(0,1));
-    ASSERT_DOUBLE_EQ(1, (double)T.get_rot()(1,1));
-    ASSERT_DOUBLE_EQ(0, (double)T.get_rot()(2,1));
+    ASSERT_DOUBLE_EQ(0, T.get_rot()(0,1));
+    ASSERT_DOUBLE_EQ(1, T.get_rot()(1,1));
+    ASSERT_DOUBLE_EQ(0, T.get_rot()(2,1));
 
-    ASSERT_DOUBLE_EQ(0, (double)T.get_rot()(0,2));
-    ASSERT_DOUBLE_EQ(0, (double)T.get_rot()(1,2));
-    ASSERT_DOUBLE_EQ(1, (double)T.get_rot()(2,2));
+    ASSERT_DOUBLE_EQ(0, T.get_rot()(0,2));
+    ASSERT_DOUBLE_EQ(0, T.get_rot()(1,2));
+    ASSERT_DOUBLE_EQ(1, T.get_rot()(2,2));
 }
 
 TEST_F(BodyTest, forced_states_are_taken_into_account)
